use finer time to avoid duplicate nonces

+++ lib/cryptopia/api/private.rb
@@ -173,7 +173,7 @@ module Cryptopia
       end
 
       def nonce
-        @nonce ||= Time.now.to_i.to_s
+        @nonce ||= Time.now.to_f.to_s
       end
 
       def invalid_transaction_type?(options)
